+++ libs/fsm/fsm.c
@@ -116,7 +116,7 @@ void am_fsm_dispatch(struct am_fsm *fsm, const struct am_event *event) {
      *  const struct am_event *e = am_event_allocate(id, size);
      *  am_event_inc_ref_cnt(e); <-- THIS IS MISSING
      *  am_fsm_dispatch(fsm, e);
-     *      am_event_push_XXX(...) & am_event_queue_pop_front_with_cb(...)
+     *      am_event_queue_push_XXX(...) & am_event_queue_pop_front_with_cb(...)
      *      OR
      *      am_event_inc_ref_cnt(e) & am_event_dec_ref_cnt(e)
      *  am_event_free(&e);